requirements.txt
fastapi==0.114.0
uvicorn[standard]==0.23.0
pydantic==2.9.2
openapi-spec-validator==0.7.1
PyYAML==6.0.2
pytest==7.2.2
pytest-asyncio==0.20.3
httpx==0.23.0